Add Park 2017 datafiles from ApJ website
I wanted to automate this so it downloaded them for maximum
transparency, but the ApJ website seems to have some anti-crawl features
that block the attempts I've made at getting the files automatically via
curl. It works occasionally, but not reliably, so I've just decided to
add them to the Git repo.